requirements.txt
Flask-Cors==5.0.0
openai>=1.6.1,<2.0.0
python-dotenv==1.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.52
langchain-openai==0.3.13
langchain-text-splitters==0.3.8
langsmith==0.1.147
PyPDF2==3.0.1
PyMuPDF==1.25.3
faiss-cpu==1.7.4
numpy<2   
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
jinja2==3.1.6
python-dateutil==2.9.0.post0
Flask==3.1.0
flask-restx==1.3.0
